+++ CHANGELOG.md
@@ -5,11 +5,16 @@
 ### Added
 
 - `Innmind\UrlTemplate\Template::attempt()`
+- `Innmind\UrlTemplate\Template::expansion()`
 
 ### Changed
 
 - Requires `innmind/immutable:~5.18`
 - Requires PHP `8.4`
+
+### Removed
+
+- `Innmind\UrlTemplate\Template::expand()`, use `::expansion()` instead
 
 ## 3.1.0 - 2023-09-16